pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ask-to-mask"
version = "0.1.0"
description = "Generate organelle segmentation masks from EM images using Flux image editing"
license = { text = "BSD-3-Clause" }
requires-python = ">=3.12"
dependencies = [
    "diffusers>=0.32.0",
    "transformers",
    "accelerate",
    "safetensors",
    "Pillow",
    "numpy",
    "scikit-image",
    "tqdm",
]

[project.optional-dependencies]
train = [
    "peft>=0.7.0",
    "zarr",
    "scipy",
    "pyyaml",
    "tensorboard",
    "bitsandbytes",
]
agents = [
    "ollama>=0.4.0",
]
agents-anthropic = [
    "ollama>=0.4.0",
    "anthropic>=0.40.0",
]
agents-google = [
    "ollama>=0.4.0",
    "google-genai>=1.0.0",
]
agents-openai = [
    "ollama>=0.4.0",
    "openai>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ask_to_mask"]
